web_monitor/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
aiortc>=1.6.0
av>=11.0.0
opencv-python>=4.8.0
numpy>=1.24
requests>=2.31
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
